# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'CineCritic API'
  version: 0.1.0
  description: 'API for fetching movie data via TMDB proxy and CineCritic resources.'
servers:
  -
    url: 'http://localhost:4000'
    description: 'Local development'
  -
    url: 'https://cinecritic.onrender.com'
    description: 'Production (Render)'
...
paths:
...
  '/api/reviews/{tmdbId}':
    get:
      summary: 'Get reviews for a movie by TMDB id'
      tags:
        - Reviews
      parameters:
        -
          in: path
          name: tmdbId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'List of reviews for the movie'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
        '500':
          $ref: '#/components/responses/ServerError'
  '/api/reviews/id/{id}':
    get:
      summary: 'Get a review by id'
      tags:
        - Reviews
      parameters:
        -
          in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'
  /api/reviews:
    post:
      summary: 'Create a review for a movie'
      tags:
        - Reviews
      security:
        -
          bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tmdbId
                - userId
                - rating
              properties:
                tmdbId:
                  type: integer
                userId:
                  type: integer
                rating:
                  type: number
                  format: float
                  description: 'Rating between 0.5 and 5 in 0.5 steps'
                body:
                  type: string
                status:
                  type: string
                  enum: [draft, published, flagged]
                  default: published
      responses:
        '201':
          description: 'Review created'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/ServerError'
  '/api/reviews/{id}':
    put:
      summary: 'Update a review'
      tags:
        - Reviews
      security:
        -
          bearerAuth: []
      parameters:
        -
          in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                rating:
                  type: number
                  format: float
                body:
                  type: string
                status:
                  type: string
                  enum: [draft, published, flagged]
      responses:
        '200':
          description: 'Review updated'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      summary: 'Delete a review'
      tags:
        - Reviews
      security:
        -
          bearerAuth: []
      parameters:
        -
          in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: 'Review deleted'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required:
        - success
        - status
        - error
        - code
        - timestamp
      properties:
        success:
          type: boolean
          example: false
        status:
          type: integer
          example: 401
        error:
          type: string
          example: Unauthorized
        code:
          type: string
          example: unauthorised
        timestamp:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00.000Z'
...
    Review:
      type: object
      required:
        - id
        - tmdbId
        - userId
        - rating
      properties:
        id:
          type: integer
          description: 'Review ID'
          example: 1
        tmdbId:
          type: integer
          description: 'TMDB movie ID'
          example: 129
        userId:
          type: integer
          description: 'User ID who wrote the review'
          example: 5
        rating:
          type: number
          format: float
          description: 'Rating between 0.5 and 5.0'
          example: 4.5
        body:
          type: string
          nullable: true
          description: 'Review text content'
          example: 'Loved the visuals and story.'
        status:
          type: string
          enum:
            - draft
            - published
            - flagged
          description: 'Review status'
          example: published
        created_at:
          type: string
          format: date-time
          nullable: true
          description: 'Review creation timestamp'
          example: '2026-01-11T07:15:00Z'
      additionalProperties: true
...
  responses:
    BadRequestError:
      description: 'Bad request'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'Invalid parameters'
            code: REQ_400
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'Missing or invalid token'
            code: AUTH_401
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden
            code: AUTH_403
    NotFoundError:
      description: 'Not found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'Not found'
            code: GEN_404
    ServerError:
      description: 'Server error'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: 'Server error'
            code: SRV_500